still work in progress.  I had to merge some changes against stefans changes before I can copy over the files with my recent work.
discovered a limitation (toughness to use) in CVS.  If I do a "cvs
remove" and then stefan does a "cvs commit" before I do mine, and he
edited the file I am removing, it is a HUGE PAIN IN THE BUTT to get
things happy again.  I had to revert to getting a new working copy of
the module, and doing the remove there.  It took me a while to figure
out what was actually wrong.  Only after that could I get update to be
happy in my real working directory.