IMPORTANT CHANGE: 	Trackers will now print NO message if they cannot find a tracker config file. They still print an error if they find a file and cannot read it. They now print a message if they DO find the file, telling that they read their configuration from it.
There is an improved vrpn.cfg.SAMPLE file in server_src. It gives a
full description of all the servers you can launch and their parameters.
It includes an example configuration for each.

Fixed a bug in the vrpn_Connection code that would cause the server to
go into a mode where it could not accept connections. This would happen
when the client died at a particularly bad time.

Added a new (better?) way to do flying based on joystick boxes, called
vrpn_Tracker_AnalogFly. You can specify the channel mappings and scales
within the configuration file, and you can attach it to any analog, not
just a UNC joystick. Also includes a "reset" button to return to the
identity transform.

The Millisecond-sleep routine in vrpn_Shared.C was fixed so that it will
release the CPU rather than busywaiting.