fixed the "doesn't pause for a little while when I tell it to" bug
(really, this bug could cause a steady stream of messages any time you
slowed down the replay rate.  pausing is the extreme of slowing down
replay rate, so that's where this bug is most evident)


One side-effect of the bug this commit fixes, is that a single call to
vrpn_FileConnection::mainloop would never play back more than two (for
our application, at least) messages.  That is why a queue of
messages-to-play was building up.

To determine if this bug affected your app, you have to consider if the
frequency at which you call mainloop, multiplied by your replay rate,
is close to or more than the frequency of messages in the file you are
playing back.  If it is more than in the file, then you were at risk
and should now check to make sure your app didn't break.  (the number
of messages that can be played by one call to mainloop is no longer
bounded to two.)


the cause of this bug was some incorrect struct timeval math.  There
is a comment on vrpn_TimevalSum (ONLY IN THE .C FILE) that says that
the two arguments must be normalized prior to calling the funct.  The
same applies to vrpn_TimevalDiff, although there is no comment there.

THERE IS NO COMMENT IN THE .H FILE!!!!!!

normalized means that abs(tv_usec) is < 1000000

However, vrpn_TimevalScale (which multiplies a timeval by a double),
does not normalize after doing the multiply.  So, the tv_usec part
would be > 1000000 if you set the replay rate high enough.  This was
causing vrpn_TimevalSum to do the wrong thing in this code

    const timeval skip_time          // scale it by d_rate
        = vrpn_TimevalScale( real_elapsed_time, d_rate ) );

So, the behavior I described (of time not advancing correctly in
vrpn_FileConnection) was really happening, but for a different reason
than I thought.

Instead of making vrpn_TimevalScale normalize (don't know if any code
depends on it's current behavior), I added a function
vrpn_TimevalNormalize, and use it to filter the result of
vrpn_TimevalScale.  The previous example is now

    const timeval skip_time          // scale it by d_rate
        = vrpn_TimevalNormalize( vrpn_TimevalScale( real_elapsed_time,
                                                    d_rate ) );

TODO:

vrpn_TimevalNormalize is not really tested.  It should be when I'm not
so tired.  (in particular, for negative values) There are still no
comments explaining these gotchas in vrpn_Shared.h.  Need to add them,
after deciding the right way to handle all this.