Changes vrpn_FileConnection so it will allow local logging. I used this ability to write a log-file manipulator. I read in a log file, check and apply a fix to the timestamps, then send out an identical message to the one I received on the file connection, which gets logged. Problem: I don't get system messages in the new log file.
Played with gettimeofday in Cygwin environment, and found it is broken and
can't be fixed using Han's tricks in vrpn_Shared.C. But, instead found a
way to get rid of vrpn_cygwin_hack.h - just include sys/time.h AFTER the
standard windows headers. vrpn_cygwin_hack.h will be totally removed in my
next commit.

Added quick hack to make Clock messages system messages. This allows clock
sync to happen before any user messages are send, if you spin the
connection's mainloop a few times on each end, and allows the
File_Connection to replay a log of the session correctly later. This is
the work-around for the broken Cygwin gettimeofday.