This version of vrpn_Connection fixes (some of?) the bugs that I introduced into the type & sender lookup when we went to multiconnection. This fix keeps the lookups linear-time but is a little hackish;  the unhackish fix is ugly, and the pretty/safe fix is not linear-time;  even though vrpn isn't usually time-critical I want to keep type and sender lookups linear time since some of those tables are growing pretty big (>200 for the microscope, >100 for nano collaboration).
This version of mutex has been tested for three peers, and works
(yay for multiconnection servers!) with one caveat:  if multiple peers
are on the same host, they could possibly deadlock (current tiebreaking
assumes unique IP numbers).  The fix is to use IP numbers + port numbers,
which are unique within a protocol family.