Fixed loop conditions in vrpn_RedundantTransmission so that we shouldn't get the crashes Adam saw.  They never occurred in my tests on Unix, but that's one thing PCs are good for - being fragile.
Fixed vrpn_DelayedConnection & wrote a new variant,
vrpn_RelativeDelayConnection.  Gave both more features (choosing to delay
only some message types rather than all).  DelayedConnection makes delayed
messages wait some fixed time after they're received before their callbacks
are triggered.  RelativeDelayConnection waits some fixed time after their
timestamp, - RTT.  If the server timestamps responses with the time the
corresponding requests were sent, this lets us attempt to give a constant
latency connection.  A later revision can attempt to integrate this
into vrpn_Connection's vrpn_CONNECTION_FIXED_LATENCY type (if the
connection weren't Synchronized we wouldn't have to subtract RTT and
it'd be more reasonable to consider it constant latency.)

In order to fix all this, I had to mess with constructors:  I'd set up
allocateEndpoint() as a virtual function so that it can be overridden in
derived classes, but it needs to be called during the constructor, at a
time when the derived class vtable isn't set up yet.  So instead, it's a
static function that is passed as a parameter to the constructor & has
a pointer stashed - manual vtables.  The one alternative that I'm aware
of or that I can find any writeups on is to have every user of
vrpn_Connection call an init() function immediately after newing it &
move most of the contents of the constructor there.  That's what we're
doing with vrpn_BaseClass, IMHO but vrpn_Connection ought to be a bit
more user-friendly than vrpn_BaseClass.  As it is users shouldn't
have to worry about these extra parameters, but we could move them onto
protected constructors if we didn't mind having 4 Synchronized & Deferred
constructors instead of 2.

	* vrpn_Connection.C (vrpn_Connection) : add data members
        d_endpointAllocator, d_updateEndpoint
	(vrpn_Connection client constructor) : add endpoint allocator
	as extra argument, call (*d_endpointAllocator) & set d_updateEndpoint
	instead of calling allocateEndpoint()
	(vrpn_Connection server constructor) : same
	(vrpn_Synchronized_Connection client constructor) : add endpoint
	allocator as extra argument
	(vrpn_Synchronized_Connection server constructor) : same
	(vrpn_Connection::updateEndpoints) : new function
	(vrpn_Connection::allocateEndpoint) : change from virtual to static
        (vrpn_Connection::connect_to_client) : replace call to
        allocateEndpoint() with (*d_endpointAllocator), set d_updateEndpoint
	(vrpn_Connection::server_check_for_incoming_connections) : same
	(vrpn_Connection::mainloop) : call updateEndpoints() if
	d_updateEndpoint is true
	(vrpn_Endpoint::dispatch) : make virtual

	* vrpn_DelayedConnection.C (vrpn_DelayedEndpoint) : new data
	members d_delayListEnd, d_delayAllTypes, d_typeIsDelayed
	(vrpn_DelayedConnection) : new data members d_delayAllTypes,
	d_typeIsDelayed
	(vrpn_DelayedEndpoint constructor) : remove timeval parameter,
	initialize d_typeIsDelayed
	(vrpn_DelayedEndpoint::delayAllTypes) : new function
	(vrpn_DelayedEndpoint::delayType) : same
	(vrpn_DelayedEndpoint::enqueue) : same
	(vrpn_DelayedEndpoint::updateEndpoints) : same
	(vrpn_DelayedEndpoint::dispatch) : don't delay system messages;
	TCP/UDP connections will timeout otherwise
	(vrpn_DelayedConnection constructor) : initialize d_typeIsDelayed
	(vrpn_DelayedConnection::delayAllTypes) : new function
	(vrpn_DelayedConnection::delayType) : same
	(vrpn_RelativeDelayEndpoint) : new class
	(vrpn_RelativeDelayConnection) : new class

	* vrpn_RedundantTransmission.C
	(vrpn_RedundantTransmission::mainloop) : detect & clean-up after
	error states