Improvements to the makefile build system. Originally 7 modifications: Clean up makefiles by splitting lines.
Make python bindings able to compile on 64-bit linux

Use provided CFLAGS and CXXFLAGS in addition to the ones we guess

Fix python makefile by not hardcoding the python version as well as
making sure the target directory exists before we build into it.

Add an install target to the quatlib makefile

Ensure all "mkdir" calls are mkdir -p to create the whole path and not error

Update the java bindings makefile:
The java makefile shares a lot with the other makefiles: selectively merge improvements
in those files into the java makefile.