revamp build system for VRPN-ICMS requirements
VRPN_SOURCE_DIR variable is now VRPN-ICMS_SOURCE_DIR
manually add libusb-1.0 and -lrt when building with libusb
(may break windows compatibility for now).
libnifalcon is no longer flagged as advanced,
since the whole point of this package is to add support
for it and nothing else will typicall be compiled in.