Re-enable directshow video server for vc10.
Apparently the internet was exaggerating - as long as you've got
an older SDK with qedit in it sitting around and you set up includes
right, (so grabbing just that header from the old sdk) it works fine.

Tested with VS10 on a machine that also has VS9 (2008) installed:
the Windows SDK v6.0A included with VS9 has qedit.h which it used.