Ensure null-termination (cppcheck warning)
"The use of strncpy() usually indicates that
the programmer wants to ensure the buffer
is zero-terminated after the call. However
if the (buffer) size given for the strncpy()
call matches the actual buffer size strncpy()
does not add the zero at the end of the buffer.
This may cause bugs later in the code if the
code assumes buffer is zero-terminated."