Big CMake changes to handle finding python.
We search for any version if no particular setting is enabled.
If SWIG or hand-coded 2x is enabled, search just for 2.
If hand-coded 3x is enabled, search just for 3.

option_requires will still enable things by default if the
dependencies (whether python2 or python3) are found when running
with a clean cache.

If any of the Python-related options are changed, re-do the
version selection and library search described above.