Fix various const issues
Fix some functions that take mutable pointers for no apparent reason to
take const pointers instead, which allows us to avoid a number of casts
to remove const (some of which are illegal, i.e. casts of string
literals).  Remove some other casts that just don't need to be done.