Disable VRPN_EXPORT_GETTIMEOFDAY by default, and fix related issues.
In the CMake build, that option is now only available on Windows, non-Cygwin,
and it's off by default. (The configure header was deceptive - it did have
an effect, mangling some Boost headers included after a VRPN header when
built on MinGW64, and anyway, it's generally impolite to pollute the
global namespace, much less a standard POSIX name (?).

That exposed a bug in the preprocessor spaghetti that was resulting in
undefined struct timezone in the case of undefined VRPN_EXPORT_GETTIMEOFDAY,
also fixed.