cmake: set the soversion on the libraries
This creates libvrpn.so.0 libraries which are typically used as the file
with the actual code. The libvrpn.so is then a symlink which is used in
development files so that -lvrpn works. Same for the server library.

This number should be bumped for ABI breaks (removal of old functions,
symbol renames, etc.).

Not required for the plain-Make build since it only builds static
libraries.