ENH: Improve behavior LibNiFalcon init fails
1) Do not close and reopen device if firware can not
be loaded: used to be done in an example of libnifalcon
but this ended up not being a good solution. Trying to load
multiple time in a row though would usually work as long
as the device stays open

2) Exit with error in init() if firmware hasn't been loaded after
the max number of attempts. Used to keep going and say firware
successfully loaded.

3) Do not reattemp a reset in case it failed, this was triggering
and infinite loop. We already tried multiple times opening the
device and loading the firware in reset(), if this failed there
is no point to automatically doing it again and again.